feat: migrate users/boards/tasks modules to TS
