+++ src/app.js
@@ -3,9 +3,9 @@ const express = require('express');
 const swaggerUI = require('swagger-ui-express');
 const path = require('path');
 const YAML = require('yamljs');
-const userRouter = require('./resources/users/user.router');
-const boardRouter = require('./resources/boards/board.router');
-const taskRouter = require('./resources/tasks/task.router');
+const userRouter = require('src/resources/users/user.router');
+const boardRouter = require('src/resources/boards/board.router');
+const taskRouter = require('src/resources/tasks/task.router');
 // import express from 'express';
 // import swaggerUI from 'swagger-ui-express';
 // import path from 'path';
