+++ src/resources/users/user.model.ts
@@ -1,6 +1,15 @@
 import {v4 as uuid} from 'uuid';
+import { IUser } from '../../types/interfaces';
 
-class User {
+class User implements IUser{
+  public login: string;
+
+  public id: string;
+
+  public name: string;
+
+  public password: string;
+
   constructor({
     id = uuid(),
     login = 'user',
@@ -13,7 +22,7 @@ class User {
     this.password = password;
   }
 
-  static toResponse(user) {
+  static toResponse(user: IUser) {
     const { id, name, login } = user;
     return { id, name, login };
   }
